common: Use int instead of long for key object ids

+++ src/main/java/com/hoccer/talk/model/TalkKey.java
@@ -20,7 +20,7 @@ public class TalkKey {
     private String _id;
 
     @DatabaseField(generatedId = true)
-    private long publicKeyId;
+    private int publicKeyId;
 
     @DatabaseField
     private String clientId;
