+++ src/main/java/com/hoccer/talk/model/TalkPrivateKey.java
@@ -14,7 +14,7 @@ public class TalkPrivateKey {
     public static final String FIELD_TIMESTAMP = "timestamp";
 
     @DatabaseField(generatedId = true)
-    private long privateKeyId;
+    private int privateKeyId;
 
     @DatabaseField(columnName = FIELD_CLIENT_ID)
     private String clientId;
